% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dive_tidy_functions.R
\name{add_dates_times}
\alias{add_dates_times}
\title{Add POSIXct date times and create a regular time series}
\usage{
add_dates_times(trace, start_time, on_seal, off_seal)
}
\arguments{
\item{trace}{tidy trace data frame after arc removal, contains the x and y
values of the trace.}

\item{start_time}{time TDR was turned on, in y:m:d h:m:s format.}

\item{on_seal}{time TDR was placed on the seal, in y:m:d h:m:s format.}

\item{off_seal}{time TDR taken off the seal, in y:m:d h:m:s format.}
}
\value{
trace data frame with POSIXct date times and interpolated points to
fill sparse parts of the record.
}
\description{
The seal often moved faster than the transducer arm could document the dive,
which results in large gaps as the seal is descending/ascending. This
function transforms the record into a regular time series, which greatly
benefits future spline smoothing and dive analysis. This also transforms
the x-axis into POSIXct date times.
}
\examples{
\dontrun{
trace <- add_dates_times(trace, start_time = "1981:01:16 15:10:00", on_seal = "1981:01:16 17:58:00", off_seal = "1981:01:23 15:30:00")
}

}
