% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\docType{data}
\name{args}
\alias{args}
\title{Arguments to use for example record}
\format{
A data frame with 13 variables
\describe{
\item{radius}{radius of transducer arm, scaled to paper records}
\item{center_y}{height of pivot point transducer arm from depth = 0}
\item{center_along_y}{y-value to center timing dots along}
\item{time_period_min}{minutes each time dot represents}
\item{spar_h}{high resolution spar value for spline smoothing}
\item{depth_thresh}{shallowest depth value for dive detection}
\item{date_start}{time TDR was turned on, in ymd_hms}
\item{max_depth}{maximum depth of paper record, if psi calibration curve is absent}
\item{k_h}{window to use for zero-offset correction}
\item{depth_bounds_l}{lower value of depth = 0, for zero-offset correction}
\item{depth_bounds_h}{higher value of depth = 0, for zero-offset correction}
\item{on_seal}{time TDR was placed on the seal}
\item{off_seal}{time TDR was taken off the seal}
}
}
\usage{
args
}
\description{
Source: Image Processing record WS_25 from Castellini et al., 1992
}
\author{
EmmaLi Tsai
}
\keyword{datasets}
