% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centering_functions.R
\name{center_scan}
\alias{center_scan}
\title{Center the data from scan misalignment}
\usage{
center_scan(trace, time_dots, center_along_y = 1.1)
}
\arguments{
\item{trace}{data frame containing the xy positions of the dive trace}

\item{time_dots}{data frame contains the xy positions of the timing dots}

\item{center_along_y}{the horizontal line to center the timing dots along}
}
\value{
Centered trace data frame
}
\description{
The record often slightly moves while being fed into the scanner, which often
results in unavoidable misalignment throughout the record. To ensure that any
drift is from the TDR and not from scanning, centering the record is
recommended as future recovery calculations assume that depth = 0 is the same
as y = 0.
}
\examples{
\dontrun{
# If you'd like to center the timing dots below the record
# along y = -0.9
center_scan(trace, time_dots, center_along_y = 0.9)
}
}
