% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.best_spar}
\alias{.best_spar}
\title{Find the best spar value}
\usage{
.best_spar(dive_stats)
}
\arguments{
\item{dive_stats}{dive statistics data frame after tidying}
}
\value{
a data frame containing the spar value that minimized overall bottom
distance and the best spar value for that record (i.e., spar value with
minimum average bottom distance / 2)
}
\description{
Find the best spar value
}
\examples{
\dontrun{
.best_spar(dive_stats)
}
}
