% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centering_functions.R
\name{.centered_psi_calibration}
\alias{.centered_psi_calibration}
\title{Centered PSI calibration curve}
\usage{
.centered_psi_calibration(trace, psi_interval = c(100, 200, 400, 600, 800))
}
\arguments{
\item{trace}{data frame containing the xy positions of the dive trace}

\item{psi_interval}{psi readings for the calibration curve, i.e., (100, 200, 400, 600, 800)}
}
\value{
data frame containing centered psi calibration curve for future calculations
}
\description{
Centered PSI calibration curve
}
\examples{
\dontrun{
.centered_psi_calibration(trace, psi_interval = c(100, 200, 400, 600, 800))
}
}
