% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.clean_divestats}
\alias{.clean_divestats}
\title{Tidy dive stats}
\usage{
.clean_divestats(dive_stats)
}
\arguments{
\item{dive_stats}{data frame containing dive stats from diveMove package}
}
\value{
tidy data frame with dive numbers and better string management
}
\description{
Tidy dive stats
}
\examples{
\dontrun{
.clean_divestats(dive_stats)
}

}
