% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dive_tidy_functions.R
\name{.create_regular_ts}
\alias{.create_regular_ts}
\title{creating regular time series for trace data}
\usage{
.create_regular_ts(trace, on_seal, off_seal)
}
\arguments{
\item{trace}{tidy trace data frame after arc removal, contains the x and y values of the trace.}

\item{on_seal}{time TDR was placed on the seal.}

\item{off_seal}{time TDR taken off the seal.}
}
\value{
trace data frame with POSIXct date times and interpolated points to fill sparse parts of the record.
}
\description{
creating regular time series for trace data
}
\examples{
\dontrun{
trace <- create_regular_ts(trace, on_seal = "1981:01:16 17:58:00", off_seal = "1981:01:23 15:30:00")
}
}
