% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.find_center_y_nopsi}
\alias{.find_center_y_nopsi}
\title{Find Center Y, without psi calibration curve}
\usage{
.find_center_y_nopsi(x1, y1, x2, y2, r = 21.14, rate, max_depth, df)
}
\arguments{
\item{x1}{x-value of the surface point}

\item{y1}{y-value of the surface point, should be close if not 0.}

\item{x2}{x-value of the point at depth}

\item{y2}{y-value of the point at depth}

\item{rate}{rate of film movement, estimated by timing dots}

\item{max_depth}{maximum depth, if known}

\item{trace}{tidy trace data frame}
}
\value{
numeric value of an estimated center_y value to use for arc removal.
}
\description{
Find an estimated value to use for the height of the transducer arm's pivot
point if only a max depth is known
}
\examples{
\dontrun{
.find_center_y_nopsi(x1, y1, x2, y2, rate, max_depth, trace)

.find_center_y_nopsi(x1 = 65.258, y1 = -0.056, x2 = 63.442, y2 = 5.341, rate = 0.21, max_depth = 484, trace)
}

}
