% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.fun}
\alias{.fun}
\title{Internal function for assigning dive numbers}
\usage{
.fun(x)
}
\arguments{
\item{x}{helper data frame for assigning dive numbers}
}
\value{
vector of dive numbers
}
\description{
Internal function for assigning dive numbers
}
\examples{
\dontrun{
.fun(helper_data_frame)
}
}
