% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.get_divestats}
\alias{.get_divestats}
\title{Get the dive statistics for each spar value iteration}
\usage{
.get_divestats(folder = "results/WS_25_1981/")
}
\arguments{
\item{folder}{folder containing the trace, time dots, and argument files.}
}
\value{
data frame of all dive statitics for each dive for all spar value
scenarios.
}
\description{
Get the dive statistics for each spar value iteration
}
\examples{
\dontrun{
filepath <- system.file("extdata", "WS_25_1981", package = "recoverKBTDR")
.get_divestats(filepath)
}
}
