% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dive_tidy_functions.R
\name{.transform_psitodepth}
\alias{.transform_psitodepth}
\title{Transform the y-axis from position to depth in meters using the psi
calibration curve}
\usage{
.transform_psitodepth(
  trace,
  psi_calibration,
  max_psi = 900,
  max_position = 22.45
)
}
\arguments{
\item{trace}{tidy trace data frame after arc removal, contains the x and y
values of the trace.}

\item{psi_calibration}{data frame containing the centered psi calibration
curve.}

\item{max_psi}{maximum psi of the TDR, often not captured in psi calibration
curve.}

\item{max_position}{position of maximum psi reading for TDR in cm.}
}
\value{
trace data frame with depth in meters.
}
\description{
Transform the y-axis from position to depth in meters using the psi
calibration curve
}
\examples{
\dontrun{
trace <- .transform_psitodepth(trace, psi_calibration, max_psi = 900, max_position = 22.45)
}
}
