% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{find_center_y_psi}
\alias{find_center_y_psi}
\title{Find Center Y, with psi calibration curve}
\usage{
find_center_y_psi(
  x1,
  y1,
  x2,
  y2,
  r = 21.14,
  rate,
  psi_calibration = psi_calibration
)
}
\arguments{
\item{x1}{x-value of the surface point}

\item{y1}{y-value of the surface point, should be close if not 0.}

\item{x2}{x-value of the point at depth}

\item{y2}{y-value of the point at depth}

\item{r}{length of transducer arm, 21.14cm}

\item{rate}{rate of film movement, estimated by timing dots}

\item{psi_calibration}{psi calibration data frame after centering}
}
\value{
numeric value of an estimated center_y value to use for arc removal.
}
\description{
Find an estimated value to use for the height of the transducer arm's pivot
point if a psi calibration curve is present.
}
\examples{
\dontrun{
find_center_y_psi(x1, y1, x2, y2, r = 21.14, rate, psi_calibration = psi_calibration)
}
}
