% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dive_tidy_functions.R
\name{remove_arc}
\alias{remove_arc}
\title{Remove left leaning arc in KBTDR records}
\usage{
remove_arc(trace, center_y)
}
\arguments{
\item{trace}{tidy trace data frame after centering, contains the x and y
values of the trace.}

\item{center_y}{height of transducer arm pivot point. This value is usually
close to 11 cm, but there is slight variation (<1 mm) at the scale of the
KBTDR. Two functions (find_center_y_psi, find_center_y_nopsi) can be used to
estimate this value.}
}
\value{
trace data frame after arc removal.
}
\description{
This function removes the characteristic left-leaning arc in the record by
using the equation of the circle the KBTDR arm makes.
}
\examples{
\dontrun{
# if the height of the transducer arm pivot point is 11.1 cm above depth = 0:
remove_arc(trace, time_dots, center_y = 11.1)
}

}
