% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centering_functions.R
\name{rollmean}
\alias{rollmean}
\title{Rolling mean}
\usage{
rollmean(x, n)
}
\arguments{
\item{x}{x-values of the timing dots}

\item{n}{window size}
}
\value{
numeric vector of rolling means
}
\description{
Rolling mean
}
\examples{
\dontrun{
rollmean(time_dots$x_val, 2)
}
}
