% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{spar_dive_stats}
\alias{spar_dive_stats}
\title{Find the best spar value for spline smoothing using dive statistics}
\usage{
spar_dive_stats(filepath = "../sample_data/WS_25_1981")
}
\arguments{
\item{filepath}{folder containing the trace, time dots, and argument files.}
}
\value{
folder in /results that contains a csv for all spar scenarios, a
dive_stats data frame containing all dive stats for each spar scenario to the
global environment, and the best spar value to use for this record.
}
\description{
Find the best spar value for spline smoothing using dive statistics
}
\examples{
\dontrun{
filepath <- system.file("extdata", "WS_25_1981", package = "recoverKBTDR")
spar_dive_stats(filepath)
}
}
