% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_trace.R
\name{tidy_raw_trace}
\alias{tidy_raw_trace}
\title{Correct raw trace csv file from ImageJ}
\usage{
tidy_raw_trace(filepath_trace, col_x = 1, col_y = 2)
}
\arguments{
\item{filepath_trace}{the path to the csv file containing the raw trace data
from ImageJ}

\item{col_x}{column of the x values, default is set to first column.}

\item{col_y}{column of the y values, default is set to second column.}
}
\value{
One trace data frame after correcting ImageJ's default origin
placement
}
\description{
This function takes the raw trace csv file from ImageJ and applies a series
of corrections to it such that it can be used in this package. This
involves correcting the default y-axis values, ordering, filtering, and
adding proper names to the data frame.
}
\examples{
\dontrun{
filepath <- system.file("extdata", "WS_25_1981", package = "recoverKBTDR")
filepath_trace <- paste(filepath, "WS_25_1981_trace.csv", sep = "/")
tidy_raw_trace(filepath_trace, col_x = 1, col_y = 2)
}

}
