% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_dots.R
\docType{data}
\name{time_dots}
\alias{time_dots}
\title{X and Y values of timing dots for a sample record}
\format{
A data frame with 2 variables
\describe{
\item{x_val}{x-value position of the point, in centimeters from the origin}
\item{y_val}{y-value position of the point, in centimeters from the origin}
}
}
\usage{
time_dots
}
\description{
Source: Image Processing record WS_25 from Castellini et al., 1992
}
\author{
EmmaLi Tsai
}
\keyword{datasets}
