% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dive_tidy_functions.R
\name{transform_x_vals}
\alias{transform_x_vals}
\title{Remove arc in record and transform x-axis to time}
\usage{
transform_x_vals(trace, time_dots, center_y = 11.1, time_period_min = 12)
}
\arguments{
\item{trace}{tidy trace data frame after centering, contains the x and y
values of the trace.}

\item{time_dots}{tidy time dots data frame, contains the x and y positions
of the timing dots.}

\item{center_y}{height of transducer arm pivot point. This value is usually
close to 11 cm, but there is slight variation (<1 mm) at the scale of the
KBTDR. Two functions (find_center_y_psi, find_center_y_nopsi) can be used to
estimate this value.}

\item{time_period_min}{minutes elapsed between two time periods.}
}
\value{
trace data frame after arc removal.
}
\description{
This function removes the characteristic left-leaning arc in the record by
using the equation of the circle the KBTDR arm makes. It also uses the timing
dots the transform the x-axis to time, in minutes from the start. This can
later be transformed to POSIXct date times and transformed into a regular
time series in the add_dates_times function.
}
\examples{
\dontrun{
# if the height of the transducer arm pivot point is 11.1 cm above depth = 0:
transform_x_vals(trace, time_dots, center_y = 11.1, time_period_min = 12)
}

}
