% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dive_tidy_functions.R
\name{transform_y_vals}
\alias{transform_y_vals}
\title{Transform the y-axis from position to depth in meters}
\usage{
transform_y_vals(
  trace,
  maxdep = NULL,
  psi_calibration = NULL,
  max_psi = NULL,
  max_position = NULL
)
}
\arguments{
\item{trace}{tidy trace data frame after arc removal, contains the x and y values of the trace.}

\item{psi_calibration}{data frame containing the centered psi calibration curve.}

\item{max_psi}{maximum psi of the TDR, often not captured in psi calibration curve.}

\item{max_position}{position of maximum psi reading for TDR in cm.}

\item{max_depth}{maximum depth of trace in meters, if psi calibration curve is not present.}
}
\value{
trace data frame with depth in meters.
}
\description{
Transforms the y-axis of the record from position to depth, using either
the psi calibration curve at the end of the record, or a known maximum depth
for that record.
}
\examples{
\dontrun{
# if the record has a psi calibration curve at the end:
trace <- transform_psitodepth(trace, psi_calibration, max_psi = 900, max_position = 22.45)

# if only the maximum depth is known:
trace <- trace(trace, max_deph = 317)
}
}
