% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centering_functions.R
\name{center_scan}
\alias{center_scan}
\title{Center the data from scan misalignment}
\usage{
center_scan(trace, time_dots, dist_timedot = 1.1)
}
\arguments{
\item{trace}{data frame containing the xy positions of the dive trace}

\item{time_dots}{data frame contains the xy positions of the timing dots}

\item{dist_timedot}{the horizontal line to center the timing dots along}
}
\value{
Centered trace data frame
}
\description{
Center the data from scan misalignment
}
\examples{
\dontrun{
center_scan(trace, time_dots, dist_timedot = 0.9)
}
}
