% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centering_functions.R
\name{center_scan}
\alias{center_scan}
\title{Center the data from scan misalignment}
\usage{
center_scan(trace, time_dots, center_along_y = 1.1, psi_interval = NULL)
}
\arguments{
\item{trace}{data frame containing the xy positions of the dive trace}

\item{time_dots}{data frame contains the xy positions of the timing dots}

\item{center_along_y}{the horizontal line to center the timing dots along}

\item{psi_interval}{optional numeric vector of psi intervals at the end of the record, if present.}
}
\value{
Centered trace data frame and centered psi_calibration curve at the end of the record (printed to global environment).
}
\description{
The record often slightly moves while being fed into the scanner, which often
results in unavoidable misalignment throughout the record. To ensure that any
drift is from the TDR and not from scanning, centering the record is
recommended as future recovery calculations assume that depth = 0 is the same
as y = 0.
}
\examples{
\dontrun{
# if the record has a psi calibration curve at the end:
center_scan(trace, time_dots, center_along_y = 0.9, psi_interval = c(100, 200, 400, 600, 800))

# if the record does not have a psi calibration curve:
center_scan(trace, time_dots, center_along_y = 0.9)
}
}
