% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_recovery.R
\name{fast_recovery}
\alias{fast_recovery}
\title{Fast recovery of record using argument file}
\usage{
fast_recovery(filepath = "../data/WS_folder")
}
\arguments{
\item{filepath}{file path of folder containing at least 3 files (trace, time dots, args). The argument file should contain columns for:
\itemize{
\item radius  : the length of the radius arm in cm. This is constant across all
records, but some of the earlier records that I have were
magnified by a slightly larger amount, and therefore the scale
is slightly different.
\item center_y: height of the pivot point of the transducer arm from depth = 0
in cm. In other words, the y-value of the center of the circle
that draws the dive record. This value may change across
records, but helper functions are available to provide
estimates (?find_center_y).
\item center_along_y: distance in cm used for centering the record. All time
dots will be centered along y = -center_along_y
\item time_period_min: time elapsed between two time dots.
\item spar_h  : spar value used for high depths of the records to increase
resolution for the smooth_trace_dive() function.
\item depth_thresh: depth threshold to use for the rolling mean function
to determine what depths should be considered diving
behavior.
\item date_start: start of the record in y:m:d h:m:s format.
\item max_depth: maximum depth to use for depth transformation. Only for records
without a psi calibration curve.
\item k_h     : larger window to use for zoc() function, if needed.
\item depth_bounds_l, depth_bounds_h: low and high depth bounds that encapsulate
where depth = 0 is likely to be. Used for zoc() function.
\item on_seal : time the tdr was placed on seal. In y:m:d h:m:s format.
\item off_seal: time the tdr was taken off seal. In y:m:d h:m:s format.
}}
}
\value{
trace data frame after full recovery, added to global environment.
}
\description{
Recover a single record fast by using an argument csv file that passes
arguments to different functions for recovery.
}
\examples{
\dontrun{
filepath <- system.file("extdata", "WS_25_1981", package = "recoverKBTDR")
fast_recovery(filepath)
}

}
