% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{find_best_spar}
\alias{find_best_spar}
\title{Find the best spar value for spline smoothing using dive statistics}
\usage{
find_best_spar(filepath = "../data/WS_folder")
}
\arguments{
\item{filepath}{folder containing the trace, time dots, and argument files.}
}
\value{
folder in /results that contains a csv for all spar scenarios, a
dive_stats data frame containing all dive stats for each spar scenario to the
global environment, and the best spar value to use for this record.
}
\description{
Function recovers a record 21 separate times, between 0 and 1 with increments
of 0.05 and assess how each change in spar value influences the dive
statistics gathered by the record. It also finds the ideal spar value for
that record by assessing changes in bottom distance. This function will take
a while to run because it is recovering the same record multiple times,
calibrating it, and retrieving the dive statistics. However, I believe this
is the best method for finding the best spar value.
}
\examples{
\dontrun{
filepath <- system.file("extdata", "WS_25_1981", package = "recoverKBTDR")
find_best_spar(filepath)
}
}
