% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_trace.R
\name{read_trace}
\alias{read_trace}
\title{Correct raw time dots and trace csv files from ImageJ}
\usage{
read_trace(filepath = "data/WS_folder")
}
\arguments{
\item{filepath}{the path to the folder containing both csv files, time dots
and trace, for a single record.}
}
\value{
Two data frames of the trace and timing dots after correcting
ImageJ's default origin placement are added to the global environment
}
\description{
This function takes the raw time dots and trace csv files from ImageJ and
applies a series of corrections to them such that they can be used in this
package. This involves correcting the default y-axis values, ordering,
filtering, and adding proper names to the data frames. This function adds
the two data frames to the global environment.
}
\examples{
\dontrun{
filepath <- system.file("extdata", "WS_25_1981", package = "recoverKBTDR")
read_trace(filepath)
}

}
