% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_trace.R
\name{read_trace}
\alias{read_trace}
\title{Correct time dots and trace csv files.}
\usage{
read_trace(filepath = "data/folder")
}
\arguments{
\item{filepath}{the path to the folder containing both csv files, time dots and trace, for a single record.}
}
\value{
Two data frames of the trace and timing dots after correcting ImageJ's default origin placement.
}
\description{
Correct time dots and trace csv files.
}
\examples{
\dontrun{
read_trace("data/folder")
}

}
