% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dive_tidy_functions.R
\name{transform_x_vals}
\alias{transform_x_vals}
\title{Remove arc in record and transform x-axis to time}
\usage{
transform_x_vals(trace, time_dots, center_y = 11.1, time_period_min = 12)
}
\arguments{
\item{trace}{tidy trace data frame after centering, contains the x and y values of the trace.}

\item{time_dots}{tidy time dots data frame, contains the x and y positions of the timing dots.}

\item{center_y}{height of transducer arm pivot point.}

\item{time_period_min}{minutes elapsed between two time periods.}
}
\value{
trace data frame after arc removal.
}
\description{
Remove arc in record and transform x-axis to time
}
\examples{
\dontrun{
transform_x_vals(trace, time_dots, center_y = 11.1, time_period_min = 12)
}

}
